🚚 Rename: 파일명 index.tsx로 통일
