+++ src/app/(main)/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import { Button } from "@/components/button/button";
-import LoadingSpinner from "@/components/spinner";
+import Button from "@/components/commons/button";
+import LoadingSpinner from "@/components/commons/spinner";
 
 export default function HomePage() {
   return (
@@ -10,20 +10,20 @@ export default function HomePage() {
       <LoadingSpinner />
       {/* Button 테스트 */}
       <div className="space-y-4 mt-6">
-        <Button variant="filled" size="sm" type="button" onClick={() => alert("Small Button Clicked!")}>
+        <Button variant="primary" size="sm" type="button" onClick={() => alert("Small Button Clicked!")}>
           Small Button
         </Button>
-        <Button variant="filled" size="md">
+        <Button variant="primary" size="md">
           Medium Button
         </Button>
-        <Button variant="filled" size="lg" className="bg-blue-500 hover:bg-blue-600">
+        <Button variant="primary" size="lg" className="bg-blue-500 hover:bg-blue-600">
           Large Button
         </Button>
-        <Button variant="filled" size="md" disabled>
+        <Button variant="primary" size="md" disabled>
           Disabled Button
         </Button>
         <div className="bg-gray-200 w-50 h-50">hi</div>
-        <Button variant="filled" size="md" fullWidth>
+        <Button variant="primary" size="md" fullWidth>
           Full-Width Button
         </Button>
       </div>
