--- src/components/spinner/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./loading-spinner";
